feat: add theme page header api
