+++ README.md
@@ -5,5 +5,5 @@
 - [x] API 모듈 작성 (Request, Response Type 정의)
 - [x] 메인 페이지 Theme 카테고리 섹션 API 구현
 - [x] 메인 페이지 실시간 급상승 선물랭킹 섹션 API 구현
-- [ ] Theme 페이지 Header API 구현
+- [x] Theme 페이지 Header API 구현
 - [ ] Theme 페이지 상품 목록 섹션 API 구현
