+++ src/components/features/Home/ThemeCategorySection/index.tsx
@@ -1,8 +1,8 @@
 import styled from "@emotion/styled";
-import { useEffect, useState } from "react";
+import { useContext } from "react";
 import { Link } from "react-router-dom";
 
-import { getThemes } from "@/api";
+import { ThemesContext } from "@/App";
 import { Container } from "@/components/common/layouts/Container";
 import { Grid } from "@/components/common/layouts/Grid";
 import { getDynamicPath } from "@/routes/path";
@@ -12,20 +12,7 @@ import type { ThemeData } from "@/types";
 import { ThemeCategoryItem } from "./ThemeCategoryItem";
 
 export const ThemeCategorySection = () => {
-  const [categories, setCategories] = useState<Array<ThemeData>>([]);
-
-  useEffect(() => {
-    const fetchThemes = async () => {
-      try {
-        const themes = await getThemes();
-        setCategories(themes);
-      } catch (error) {
-        console.error("Error fetching themes:", error);
-      }
-    };
-
-    fetchThemes();
-  }, []);
+  const themes: ThemeData[] = useContext(ThemesContext);
 
   return (
     <Wrapper>
@@ -36,7 +23,7 @@ export const ThemeCategorySection = () => {
             md: 6,
           }}
         >
-          {categories.map((element) => (
+          {themes.map((element) => (
             <Link key={element.key} to={getDynamicPath.theme(element.key)}>
               <ThemeCategoryItem image={element.imageURL} label={element.label} />
             </Link>
